planning_temp/auth_backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Validation
pydantic[email-validator]==2.5.2

# Environment
python-dotenv==1.0.0
